Use daemon storage options in builder
The builder currently only uses the daemon default storage options when creating
a layer for a RUN instruction. This change uses the daemon storage options when
creating destination layers for COPY/ADD instructions as well. It currently
does not allow overriding of the default options, and only applies to the RW
layer, not the RO commit layer.

Signed-off-by: Darren Stahl <darst@microsoft.com>